0.4 Finished All Websites
Added all of the fonts, including newsreader, space mono, etc. Did all of the different CSS color schemes, all the same CSS but different colors and images. Added all of the images, such as the logos like the arcade cabinet, and the actual images like the chocolate boxes. Made all of the buttons link to the different pages, as well as having an abbr tag that shows what website each of the buttons goes to. Did all of the HTML as well, all the same HTML as well just different class names and headers.